Rename `first-published-tag-for-merged-pr` to `closing-remarks`
