+++ source/options-storage.ts
@@ -43,6 +43,7 @@ export const renamedFeatures = new Map<string, string>([
 	['quick-review-buttons', 'one-click-review-submission'],
 	['wait-for-build', 'wait-for-checks'],
 	['pull-request-hotkey', 'pull-request-hotkeys'],
+	['first-published-tag-for-merged-pr', 'closing-remarks'],
 ]);
 
 export function getNewFeatureName(possibleFeatureName: string): FeatureID | undefined {
